+++ src/java/grails/plugin/multitenant/core/exception/TenantSecurityException.java
@@ -3,19 +3,19 @@ package grails.plugin.multitenant.core.exception;
 @SuppressWarnings("serial")
 public class TenantSecurityException extends TenantException {
 
-    private Integer currentTenantId, loadedTenantId;
+    private Long currentTenantId, loadedTenantId;
 
-    public TenantSecurityException(String message, Integer currentTenantId, Integer loadedTenantId) {
+    public TenantSecurityException(String message, Long currentTenantId, Long loadedTenantId) {
         super(message);
         this.currentTenantId = currentTenantId;
         this.loadedTenantId = loadedTenantId;
     }
 
-    public Integer getCurrentTenantId() {
+    public Long getCurrentTenantId() {
         return currentTenantId;
     }
 
-    public Integer getLoadedTenantId() {
+    public Long getLoadedTenantId() {
         return loadedTenantId;
     }
